requirements.txt
autogenstudio
PyPDF2
autogen-core
autogen-agentchat
autogen-ext[openai]
autogen-ext[magentic-one,openai,web-surfer]
autogen-ext
azure-ai-projects
azure-monitor-opentelemetry
opentelemetry-instrumentation-openai-v2
streamlit
autogen-ext[web-surfer]
azure-ai-evaluation
python-docx
openai
streamlit-mermaid
python-pptx
arxiv
playwright
PyMuPDF
marshmallow==3.18.0
streamlit-quill
azure-ai-evaluation[redteam] 